test: migrate standalone verification script to pytest
* Move `verify_conditional_data.py` into `tests/test_conditional_data.py`
* Convert standalone verification logic into `pytest` async tests
* Remove unused `sys` import from the original script
* Delete obsolete standalone script
